rbsp types

+++ H264Decoder/Data/NALU/RbspSliceTrailingBits.h
@@ -5,7 +5,7 @@
 struct RbspSliceTrailingBits
 {
     RbspTrailingBits rbspTrailingBits;
-    std::uint8_t cabacZeroWord = 0;
+    std::uint16_t cabacZeroWord = 0;
 
     RbspSliceTrailingBits() = default;
     explicit RbspSliceTrailingBits(DecodingContext& context, BitstreamReader& reader)
